waveform: Refactor timing imports
