+++ src/nitypes/waveform/_analog_waveform.py
@@ -14,9 +14,7 @@ from nitypes.waveform._extended_properties import (
     ExtendedPropertyDictionary,
 )
 from nitypes.waveform._scaling import NO_SCALING, ScaleMode
-from nitypes.waveform._timing._conversion import convert_timing
-from nitypes.waveform._timing._precision import PrecisionTiming
-from nitypes.waveform._timing._standard import Timing
+from nitypes.waveform._timing import PrecisionTiming, Timing, convert_timing
 
 if sys.version_info < (3, 10):
     import array as std_array
